Allow individual requests to provide requests kwargs
This is necessary since some settings, like the proxies to use for a
request, cannot be set at a session level when using Nectar like a
general download library. This is pretty hacky, but it seems like the
best approach to allow request-by-request configuration.

This is part of the fix for https://pulp.plan.io/issues/2111.