Change how Nectar configures requests to be thread-safe
Previously, Nectar set the configuration at a session level. This didn't
cause problems because Nectar is only ever used to download with one set
of settings. However, the pulp_streamer has to change the configuration
on a request-by-request basis _and_ it needs to use a session to pool
connections. This leads to trouble when attempting to update the
configuration since Nectar changes it for all requests rather than a
specific one.

This patch instead adds a method to translate a Nectar configuration to
a set of requests kwargs that work with the ``requests.request`` API.
This allows the streamer to stash its session instance between instances
of a Nectar HTTPThreadedDownloader without corruption.

fixes #2111